Incorporate Changes from Emmanuel Frecon
Incorporates the approach from https://github.com/efrecon/docker-images of building `stunnel` from source, so that:
 - We can use the latest secure version (5.50), which is more recent than even the `alpine:edge` package.
 - We can use PSK if we so desire.

In addition, I separated the code paths for client vs. server PKI deployment and DRY-ed up the code a bit so that it's clear that the only difference between the branches of logic is in the template being chosen.